requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
pyswisseph==2.10.3.2
geopy==2.4.1
timezonefinder==6.2.0
pytz==2023.3
python-dotenv==1.0.0

# do: pip install -r requirements.txt



python-doctr

cloudinary 
gunicorn
pytest
pytest-asyncio
openai>=1.42.0
pinecone

numpy
pandas
PyYAML
requests
httpx